Purple SE 1.981 Alpha

+++ net2/Tag.js
@@ -27,6 +27,7 @@ const fs = require('fs');
 const Promise = require('bluebird');
 const DNSMASQ = require('../extension/dnsmasq/dnsmasq.js');
 const routing = require('../extension/routing/routing.js');
+const freeradius = require('../extension/freeradius/freeradius.js');
 const dnsmasq = new DNSMASQ();
 const Monitorable = require('./Monitorable');
 const Constants = require('./Constants.js');
@@ -37,6 +38,7 @@ const HostTool = require('./HostTool.js');
 const hostTool = new HostTool();
 const fwapc = require('./fwapc.js');
 const {delay} = require('../util/util.js');
+const platform = require('../platform/PlatformLoader.js').getPlatform();
 
 const envCreatedMap = {};
 
@@ -155,6 +157,15 @@ class Tag extends Monitorable {
       log.error(`Failed to add ${Tag.getTagDeviceMacSetName(uid)} to ${Tag.getTagDeviceSetName(uid)}`, err.message);
     });
     envCreatedMap[uid] = 1;
+  }
+
+  async resetPolicies() {
+    await this.loadPolicyAsync();
+    for (const key of Object.keys(this.policy)) {
+      if (key === "freeradius_server") {
+        await freeradius.reconfigServer(this.o.uid, {});
+      }
+    }
   }
 
   async createEnv() {
@@ -486,6 +497,8 @@ class Tag extends Monitorable {
   }
 
   async fwapcSetGroupMACs() {
+    if (!platform.isFireRouterManaged())
+      return;
     const HostManager = require('./HostManager.js');
     const hostManager = new HostManager();
     const macs = await hostManager.getTagMacs(this.o.uid).then(results => results.filter(m => hostTool.isMacAddress(m))).catch((err) => {
@@ -500,6 +513,8 @@ class Tag extends Monitorable {
   }
 
   async fwapcDeleteGroup() {
+    if (!platform.isFireRouterManaged())
+      return;
     await fwapc.deleteGroup(this.o.uid).catch((err) => {
       log.error(`Failed to delete group ACL in fwapc for ${this.getTagType()} ${this.o.uid}`, err.message);
     });
